+++ source/bullet.h
@@ -58,7 +58,7 @@ public:
 	RECT view;
 
 public:
-	void init(int setCode, int setX, int setY, int setDir);
+	void init(int setCode, int setX, int setY, uint8_t setDir);
 
 	void update();
 };
